test: Explicitly define 'settings_dir' and 'module_id' parameters
The 'ParseSettingsBuilder' class defines default parameter values for
'parse_settings_dir' and 'module_id', which fit to the quant/lfq/ion/dda
module. It's better to not rely on default arguments and explicitly
define 'parse_settings_dir' and 'module_id' like it is done in the tests
for quant/lfq/ion/dia module.