refactor: Replace hard coded 'parse_settings_dir' variable
Remove the 'parse_settings_dir' variable from the 'benchmarking' method,
and retrieve the path from 'quant_module/constants.py' as it is done for
the other lfq modules. This commit is simply for consistency, although
the quant_lfq_peptidoform_DIA in general is not implemented properly.

Note that no .toml settings file exist for the quant_lfq_peptidoform_DIA
module and that the path to the settings dir is therefore also missing
from 'quant_module/constants.py'.