fix: Correct wrong path specified for 'parse_settings_dir'
In the current implementation this variable is not used, as initiatlizing
the ParseSettingsBuilder is done by using its default value for the
parse_settings_dir variable.

In order to remove the default values from the ParseSettingsBuilder, the
'parse_settings_dir' variable in the webinterface DDA/variables.py must be
fixed and used.